Added Functional Requirements
Added Functional Requirements to the system.

Also decided with  @Tymac15 and @u20612894 : We will add a page where students can view the files uploaded.

In case there was any confusion as well, institutions won't have a page with a description and pictures, but only manage their page with regards to uploads etc.

Co-Authored-By: Tyrone MacLeod <105007195+Tymac15@users.noreply.github.com>
Co-Authored-By: u20612894 <114522518+u20612894@users.noreply.github.com>