User groups (#19)
* initial project set-up, with dynamoDB,S3, Auth and GraphQL API

* removing env

* adding user groups config

* removed user Auth -- want to fully design and add functionality

* university domains for user email varification

* cleaning data source

* uni names with domains and countyr

* Student,Lecture and institution data models

* uni info for demo 1

* Adding Course

Course type has been added to schema

* Added Enrollment

The Enrollment type is used to facilitate many to many relationships between students and their courses.

* Added Course Activity

Activity types have been added with a one to many relationship. A Course has multiple Activities

* Added Assessments

Assessments differ from Activities in that Activities are reoccurring and Assessments are not. Hence an Assessment has a date and an Activity has a date

* Adding Timetable

A many to many relationship has been defined between Students and Activities representing a timetable

* Adding Lecturers

Lecturers have been added to the schema

* fixing schema path

---------

Co-authored-by: u20612894 <u20612894@tuks.co.za>