Fixed require lecturer auth
Both auth components have been fixed and I have tested the following: when not signed in, or signed into admin account, there is no access to lecturer pages (working correctly, redirects to lecturer login) and vice versa for a lecturer account. One issue I have picked up is that if a user does not explicitly sign out of their account (including after closing the app and restarting), they are still able to access the accounts pages that was signed into. This session related issue needs to be looked into, however the protected routes components are now working as intended.